# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoChat 聊天/好友/房间管理后端接口",
        "title": "GoChat Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/profiles/rooms/{room_id}": {
            "get": {
                "description": "根据房间ID获取房间公开资料",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "获取房间资料",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回房间资料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_profile_port_http.GetRoomProfileResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "房间不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定房间的资料（名称、简介等），通常需要房主或管理员权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "更新房间资料",
                "parameters": [
                    {
                        "description": "更新房间资料请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_profile_port_http.UpdateRoomProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "房间资料不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
        "/rooms/requests/{request_id}/agree": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "同意指定房间成员请求，将对方加入房间",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roomship"
                ],
                "summary": "同意入群请求",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "成员请求ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "同意成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或请求已处理/无权限",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/rooms/requests/{request_id}/refuse": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "拒绝指定房间成员请求",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roomship"
                ],
                "summary": "拒绝入群请求",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "成员请求ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拒绝成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或请求已处理/无权限",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/rooms/{room_id}/members": {
            "get": {
                "description": "根据房间ID获取房间成员列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roomship"
                ],
                "summary": "获取房间成员列表",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回房间成员列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_roomship_port_http.ListRoomMembersResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "404": {
                        "description": "房间不存在",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gochat_internal_profile_dto.RoomProfile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.RoomID"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "gochat_internal_roomship_domain.RoomshipID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
        "gochat_internal_roomship_domain.RoomshipRole": {
            "type": "string",
            "enum": [
                "owner",
                "member"
            ],
            "x-enum-varnames": [
                "OwnerRole",
                "MemberRole"
            ]
        },
...
        "gochat_internal_roomship_dto.Roomship": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_roomship_domain.RoomshipID"
                },
                "role": {
                    "$ref": "#/definitions/gochat_internal_roomship_domain.RoomshipRole"
                },
                "room_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.RoomID"
                },
                "user_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                }
            }
        },
        "gochat_internal_shared_api.Code": {
            "type": "integer",
            "enum": [
                200,
                500,
                400,
                401,
                402,
                403,
                404
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeServerError",
                "CodeTimeout",
                "CodeInvalidParam",
                "CodeInvalidToken",
                "CodeNotFound",
                "CodeServiceUnavailable"
            ]
        },
        "gochat_internal_shared_api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/gochat_internal_shared_api.Code"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "gochat_internal_shared_kernel.RoomID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
        "gochat_internal_shared_kernel.UserID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
...
        "internal_profile_port_http.GetRoomProfileResponseData": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/gochat_internal_profile_dto.RoomProfile"
                }
            }
        },
...
        "internal_profile_port_http.UpdateRoomProfileRequest": {
            "type": "object",
            "required": [
                "roomID"
            ],
            "properties": {
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "roomID": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.RoomID"
                }
            }
        },
...
        "internal_roomship_port_http.ListRoomMembersResponseData": {
            "type": "object",
            "properties": {
                "roomships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gochat_internal_roomship_dto.Roomship"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}